# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/containers/info": {
            "get": {
                "description": "It returns the allocated container's data that is matched with the API key. \u003cbr\u003eNormally, only Runtime Agent servers call this endpoint.\u003cbr\u003e If `X-API-KEY` is missing, 400 error is returned.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "containers"
                ],
                "summary": "Return the containers' data matched with the API key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key assigned to each Agent server.",
                        "name": "X-API-KEY",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/containers.MyContResp"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.CommonErrorResp"
                        }
                    }
                }
            }
        },
        "/containers/launch": {
            "post": {
                "description": "This is called when the user tries to connect or to use runtime-related functionalities. \u003cbr\u003eThis endpoint launches new Docker Container that will be allocated for the user if not allocated yet, and then returns its information in order for the user to access it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "containers"
                ],
                "summary": "Launch container if not assigned, and then return the assigned one.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's Json Web Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "example": "C gcc11",
                        "description": "target runtime name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "example": "C",
                        "description": "target language",
                        "name": "lang",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/containers.LaunchResp"
                        }
                    },
                    "400": {
                        "description": "'`name` or `lang` is missing.' | 'It is launching. Please wait for a moment.'",
                        "schema": {
                            "$ref": "#/definitions/utils.CommonErrorResp"
                        }
                    }
                }
            }
        },
        "/runtimes/available": {
            "get": {
                "description": "Return supported languages available",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runtimes"
                ],
                "summary": "Return supported languages available",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/runtimes.SupportedLangResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "containers.LaunchResp": {
            "type": "object",
            "properties": {
                "port": {
                    "type": "integer",
                    "example": 8000
                },
                "url": {
                    "type": "string",
                    "example": "127.0.0.1"
                }
            }
        },
        "containers.MyContResp": {
            "type": "object",
            "properties": {
                "cont_auth": {
                    "type": "string"
                },
                "cont_auth_type": {
                    "type": "string"
                },
                "cont_ip": {
                    "type": "string"
                },
                "cont_port": {
                    "type": "integer"
                },
                "cont_user": {
                    "type": "string"
                }
            }
        },
        "runtimes.RuntimeImage": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "language_name": {
                    "description": "ForeignKey",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "runtimes.SupportedLangResp": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/runtimes.SupportedLanguage"
                    }
                }
            }
        },
        "runtimes.SupportedLanguage": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/runtimes.RuntimeImage"
                    }
                },
                "name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                }
            }
        },
        "utils.CommonErrorResp": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}